Typewriter animation
Fixed issues with typewriter animation.

When a tab is opened, the text within the title element appears to be typed. Every time a new tab is opened all of the titles are reset to empty elements so that the animation can play again.

This could be streamlined by only resetting the tab that was just open instead of looping through every one every time but I'm unsure how to do that and since there's only 4 tabs with this animation and I don't plan to expand, this should be fine for now.